+++ src/test/util/stubtool.ts
@@ -1,4 +1,4 @@
-import * as util from '../../util/exectools';
+import * as util from '../../util/exec/exectools';
 import * as vscode from 'vscode';
