+++ src/util/path/filepathtools.ts
@@ -2,19 +2,12 @@ import * as vscode from 'vscode';
 import * as ts from 'typescript';
 import * as path from 'path';
 import * as fs from 'fs';
-import { filePathToTargetPath } from './targettools';
-import { Cache } from '../model/basecache';
-import { showErrorMessage } from '../userinteraction';
 import { homedir } from 'os';
-import { BUILD_FILE, getConfig } from '../config/config';
-
-const tsExtension = vscode.extensions.getExtension('vscode.typescript-language-features');
-let active = true; 
-    
-if (!tsExtension) {
-    showErrorMessage('TypeScript Language Features are not enabled.');
-    active = false; 
-}
+import { Cache } from '../../model/basecache';
+import { getConfig } from '../../config/config';
+import { FilesContext } from '../../model/bazelquery/filescontext';
+import { TargetInfo } from '../../model/bazelquery/targetinfo';
+import { uriToBuild } from './uritools';
 
 let configCache: Cache<string, ts.ParsedCommandLine> = new Cache<string, ts.ParsedCommandLine>(getConfig("maxCacheSize"));
 
@@ -22,7 +15,7 @@ export function resolveSpecifierToFilePath(
     hostFileUri: vscode.Uri,
     specifier: string
 ) {
-    const configPath = ts.findConfigFile(hostFileUri.fsPath, ts.sys.fileExists, 'tsconfig.json'); // This can also be used for bazel build
+    const configPath = ts.findConfigFile(hostFileUri.fsPath, ts.sys.fileExists, 'tsconfig.json');
 
     if (!configPath) {
         return undefined; 
@@ -86,55 +79,13 @@ function getConfiguration(configPath: string): ts.ParsedCommandLine | undefined 
     return parsedCommandLine;
 }
 
-export function fpToBuildTarget(fsPath: string) {
-    const fileUri = vscode.Uri.file(fsPath);
-    const [target, ] = uriToBuild(fileUri) ?? [, ];
-    return target;
-}
-
 export function fpToBuild(fsPath: string) {
     const fileUri = vscode.Uri.file(fsPath);
-    const [, buildUri] = uriToBuild(fileUri) ?? [,];
+    const buildUri = uriToBuild(fileUri);
     return buildUri; 
 }
 
-// TODO: consider removing the target logic
-export function uriToBuild(fileUri: vscode.Uri): [string, vscode.Uri] | undefined {
-    const configPath = ts.findConfigFile(path.dirname(fileUri.fsPath), ts.sys.fileExists, BUILD_FILE);
-    if (!configPath) {
-        return undefined;
-    }
-
-    const rawTarget = filePathToTargetPath(configPath);
-    const index = rawTarget?.indexOf(BUILD_FILE) ?? 1;
-    const target = rawTarget?.slice(0, index - 1); 
-    if (target === undefined) {
-        return target;
-    }
-
-    const targetUri = vscode.Uri.file(configPath); 
-
-    return [target, targetUri]; 
-}
-
-export async function getBuildTargetFromFilePath(importPath: string, currentFile: vscode.Uri): Promise<[string, vscode.Uri]> {
-    const filePath = importToFullPath(currentFile, importPath);
-    if (filePath === undefined) {
-        throw new Error("Bad file path"); 
-    }
-    const fileUri: vscode.Uri | undefined = vscode.Uri.file(filePath);
-
-    if (!fileUri) {
-        throw new Error("File not found");
-    }
-    const target = uriToBuild(fileUri);
-    if (!target) {
-        throw new Error("Target not found");
-    }
-    return target;
-};
-
-export function importToFullPath(currentFile: vscode.Uri, importPath: string) {
+export function importToFs(currentFile: vscode.Uri, importPath: string) {
     let filePath: string | undefined;
 
     if (importPath.startsWith('.')) {
@@ -165,17 +116,20 @@ export function wsToFsPath(wsPath: string): string {
     return getRoot() + relativePath;
 }
 
-export function bazelLabelToUri(src: string): vscode.Uri {
-    const wsPath = src.replace(':', '/');
-    const fsPath = wsToFsPath(wsPath);
-    return vscode.Uri.file(fsPath);
-}
-
-export function bazelLabelToUris(srcs: string[]) {
-    return srcs.map(bazelLabelToUri);
-}
-
 export function getRoot() {
     const cwd = path.join(homedir(), '/lucid/main/'); // TODO get root; 
     return cwd;
 }
+
+export function pathsToTargets(importPaths: string[], context: FilesContext<string, string, TargetInfo>): Set<string> {
+    const targets = new Set<string>();
+    for (const importPath of importPaths) {
+        const wsPath = fsToWsPath(importPath);
+        const target = context.getTarget(wsPath);
+        if (target !== undefined) {
+            targets.add(target);
+        }
+    }
+    return targets;
+}
+
